refactor: update ESLint configuration for TypeScript support; modify import paths and clean up question and category entity types
