+++ src/entities/Category/api/getCategories/getCategories.ts
@@ -4,7 +4,7 @@ import { ENDPOINT_CATEGORIES } from '../../lib/endpoints';
 import type { CategoryModel } from './getCategories.types';
 
 export const getCategories = async () => {
-  const { data } = await instance.get<CategoryModel>(ENDPOINT_CATEGORIES);
+  const { data } = await instance.get<CategoryModel[]>(ENDPOINT_CATEGORIES);
 
   return data;
 };
