+++ src/entities/Question/api/getQuestions/index.ts
@@ -1,8 +1,2 @@
 export { getQuestions } from './getQuestions';
-export type {
-  QuestionModel,
-  RequestQuestionsParams,
-  Difficulty,
-  Answers,
-  IsCorrect,
-} from './getQuestions.types';
+export type { RequestQuestionsParams } from './getQuestions.types';
