+++ src/entities/Question/index.ts
@@ -1,10 +1,4 @@
-export {
-  getQuestions,
-  type QuestionModel,
-  type RequestQuestionsParams,
-  type Difficulty,
-  type Answers,
-  type IsCorrect,
-} from './api/getQuestions';
+export { getQuestions, type RequestQuestionsParams } from './api/getQuestions';
 
 export { limitQuestions } from './model/config';
+export type { QuestionModel } from './model/types';
